+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [6.0.2] - 2023-04-16
+
+- Implemented `StringExtensions::IsEmpty` extension method.
+- Implemented `EnumerableExtensions::IsEmpty` extension method.
+
 ## [6.0.1] - 2023-04-15
 
 - Set Maybe type as readonly struct.